feat: 🎸 add NonFungible details method
moves current `asset.details` method into `BaseAsset`. `NftCollection`
wraps the base call to return `totalSupply` property.

NftCollection needs to make an additional call to the chain and iterate
all holders to calculate `totalSupply` of the Nfts, which is the number
of held NFTs in the collection.

A deferred optimization is to leverage middleware (SubQuery) when provided.
However it is expected a single collection would need hundreds of unique
holders before performance improvements become noticable.
