+++ src/api/entities/Asset/index.ts
@@ -1,10 +1,4 @@
-import { Bytes, Option, StorageKey } from '@polkadot/types';
-import {
-  PalletAssetSecurityToken,
-  PolymeshPrimitivesAgentAgentGroup,
-  PolymeshPrimitivesIdentityId,
-  PolymeshPrimitivesTicker,
-} from '@polkadot/types/lookup';
+import { Bytes } from '@polkadot/types';
 import BigNumber from 'bignumber.js';
 import { groupBy, map } from 'lodash';
 
@@ -16,7 +10,6 @@ import {
   controllerTransfer,
   Identity,
   modifyAsset,
-  PolymeshError,
   redeemTokens,
   setVenueFiltering,
   toggleFreezeTransfers,
@@ -30,7 +23,6 @@ import {
 import { Query } from '~/middleware/types';
 import {
   ControllerTransferParams,
-  ErrorCode,
   EventIdentifier,
   HistoricAgentOperation,
   HistoricAssetTransaction,
@@ -46,12 +38,8 @@ import {
 } from '~/types';
 import { Ensured } from '~/types/utils';
 import {
-  assetTypeToKnownOrId,
   balanceToBigNumber,
-  bigNumberToU32,
-  boolToBoolean,
   bytesToString,
-  identityIdToString,
   middlewareEventDetailsToEventIdentifier,
   middlewarePortfolioToPortfolio,
   stringToTicker,
@@ -172,101 +160,6 @@ export class Asset extends BaseAsset {
    * @throws if the passed values result in no changes being made to the Asset
    */
   public modify: ProcedureMethod<ModifyAssetParams, Asset>;
-
-  /**
-   * Retrieve the Asset's data
-   *
-   * @note can be subscribed to
-   */
-  public details(): Promise<AssetDetails>;
-  public details(callback: SubCallback<AssetDetails>): Promise<UnsubCallback>;
-
-  // eslint-disable-next-line require-jsdoc
-  public async details(
-    callback?: SubCallback<AssetDetails>
-  ): Promise<AssetDetails | UnsubCallback> {
-    const {
-      context: {
-        polymeshApi: {
-          query: { asset, externalAgents },
-        },
-      },
-      ticker,
-      context,
-    } = this;
-
-    const assembleResult = async (
-      optToken: Option<PalletAssetSecurityToken>,
-      agentGroups: [
-        StorageKey<[PolymeshPrimitivesTicker, PolymeshPrimitivesIdentityId]>,
-        Option<PolymeshPrimitivesAgentAgentGroup>
-      ][],
-      assetName: Option<Bytes>
-    ): Promise<AssetDetails> => {
-      const fullAgents: Identity[] = [];
-
-      if (optToken.isNone) {
-        throw new PolymeshError({
-          message: 'Asset detail information not found',
-          code: ErrorCode.DataUnavailable,
-        });
-      }
-
-      const { totalSupply, divisible, ownerDid, assetType: rawAssetType } = optToken.unwrap();
-
-      agentGroups.forEach(([storageKey, agentGroup]) => {
-        const rawAgentGroup = agentGroup.unwrap();
-        if (rawAgentGroup.isFull) {
-          fullAgents.push(new Identity({ did: identityIdToString(storageKey.args[1]) }, context));
-        }
-      });
-
-      const owner = new Identity({ did: identityIdToString(ownerDid) }, context);
-      const { value, type } = assetTypeToKnownOrId(rawAssetType);
-
-      let assetType: string;
-      if (value instanceof BigNumber) {
-        const customType = await asset.customTypes(bigNumberToU32(value, context));
-        assetType = bytesToString(customType);
-      } else {
-        assetType = value;
-      }
-
-      return {
-        assetType,
-        nonFungible: type === 'NonFungible',
-        isDivisible: boolToBoolean(divisible),
-        name: bytesToString(assetName.unwrap()),
-        owner,
-        totalSupply: balanceToBigNumber(totalSupply),
-        fullAgents,
-      };
-    };
-
-    const rawTicker = stringToTicker(ticker, context);
-
-    const groupOfAgentPromise = externalAgents.groupOfAgent.entries(rawTicker);
-    const namePromise = asset.assetNames(rawTicker);
-
-    if (callback) {
-      const groupEntries = await groupOfAgentPromise;
-      const assetName = await namePromise;
-
-      return asset.tokens(rawTicker, async securityToken => {
-        const result = await assembleResult(securityToken, groupEntries, assetName);
-
-        // eslint-disable-next-line @typescript-eslint/no-floating-promises -- callback errors should be handled by the caller
-        callback(result);
-      });
-    }
-
-    const [token, groups, name] = await Promise.all([
-      asset.tokens(rawTicker),
-      groupOfAgentPromise,
-      namePromise,
-    ]);
-    return assembleResult(token, groups, name);
-  }
 
   /**
    * Retrieve the Asset's funding round
